extract benchmark-report component
