+++ package.json
@@ -18,7 +18,6 @@
     "node": ">= 0.10.0"
   },
   "author": "Robin Ward",
-  "license": "MIT",
   "devDependencies": {
     "broccoli-asset-rev": "^2.1.2",
     "broccoli-bower": "^0.2.1",
@@ -32,6 +31,7 @@
     "ember-cli-app-version": "0.5.0",
     "ember-cli-babel": "^5.1.3",
     "ember-cli-dependency-checker": "^1.0.1",
+    "ember-cli-format-number": "1.0.2",
     "ember-cli-htmlbars": "0.7.9",
     "ember-cli-htmlbars-inline-precompile": "^0.2.0",
     "ember-cli-ic-ajax": "0.2.1",
